304. Range Sum Query 2D - Immutable
